+++ modules/graphics/builtin/cpp/Gateway.cpp
@@ -35,6 +35,7 @@
 #include "refreshBuiltin.hpp"
 #include "drawnowBuiltin.hpp"
 #include "imreadBuiltin.hpp"
+#include "imwriteBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -83,6 +84,7 @@ static const nlsGateway gateway[] = {
     { "refresh", (ptrBuiltin)Nelson::GraphicsGateway::refreshBuiltin, 0, 1, CPP_BUILTIN },
     { "drawnow", (ptrBuiltin)Nelson::GraphicsGateway::drawnowBuiltin, 0, 0, CPP_BUILTIN },
     { "imread", (ptrBuiltin)Nelson::GraphicsGateway::imreadBuiltin, -1, 1, CPP_BUILTIN },
+    { "imwrite", (ptrBuiltin)Nelson::GraphicsGateway::imwriteBuiltin, 0, -2, CPP_BUILTIN },
 };
 //=============================================================================
 static bool